rename `utils.test.ts` into `internal/stdout.test.ts` in the `log` package
