+++ packages/log/src/internal/stdout.test.ts
